refactor tests
restructure test domains
move to right packages and tests
use ARM
