+++ tinkerpop3/src/test/java/io/shiftleft/overflowdb/storage/SerializerTest.java
@@ -4,8 +4,8 @@ import io.shiftleft.overflowdb.structure.NodeFactory;
 import io.shiftleft.overflowdb.structure.NodeRef;
 import io.shiftleft.overflowdb.structure.OdbConfig;
 import io.shiftleft.overflowdb.structure.OdbGraph;
-import io.shiftleft.overflowdb.structure.OdbTestEdge;
-import io.shiftleft.overflowdb.structure.OdbTestNode;
+import io.shiftleft.overflowdb.testdomains.simple.OdbTestEdge;
+import io.shiftleft.overflowdb.testdomains.simple.OdbTestNode;
 import org.apache.tinkerpop.gremlin.structure.Direction;
 import org.apache.tinkerpop.gremlin.structure.Edge;
 import org.apache.tinkerpop.gremlin.structure.T;
