+++ tinkerpop3/src/test/java/io/shiftleft/overflowdb/testdomains/gratefuldead/GratefulDead.java
@@ -1,0 +1,28 @@
+package io.shiftleft.overflowdb.testdomains.gratefuldead;
+
+import io.shiftleft.overflowdb.structure.OdbConfig;
+import io.shiftleft.overflowdb.structure.OdbGraph;
+import org.apache.tinkerpop.gremlin.structure.io.IoCore;
+
+import java.io.IOException;
+import java.util.Arrays;
+
+public class GratefulDead {
+  public static OdbGraph newGraph() {
+    return OdbGraph.open(
+        OdbConfig.withoutOverflow(),
+        Arrays.asList(Song.factory, Artist.factory),
+        Arrays.asList(FollowedBy.factory, SungBy.factory, WrittenBy.factory)
+    );
+  }
+
+  public static OdbGraph newGraphWithData() throws IOException {
+    OdbGraph graph = newGraph();
+    loadData(graph);
+    return graph;
+  }
+
+  public static void loadData(OdbGraph graph) throws IOException {
+    graph.io(IoCore.graphml()).readGraph("src/test/resources/grateful-dead.xml");
+  }
+}
